qtdemux: fix offset calculation when parsing CENC aux info
Commit 7d7e54ce6863ff53e188d0276d2651b65082ffdb added support for
DASH common encryption, however commit
bb336840c0b0b02fa18dc4437ce0ded3d9142801 that went onto master
shortly before the CENC commit caused the calculation of the CENC
aux info offset to be incorrect.

The base_offset was being added if present, but if the base_offset
is relative to the start of the moof, the offset was being added twice.
The correct approach is to calculate the offset from the start of the
moof and use that offset when parsing the CENC aux info.