debian: remove custom parallel compilation
In order to build a debian package with upstream source, the user should
do

  ./autogen.sh
  cp -a debian.upstream debian
  debuild -eDEB_BUILD_OPTIONS="parallel=8" -us -uc -b

The environment variable DEB_BUILD_OPTIONS="parallel=8" is the canonical
way to make a parallel build (-j8 in this case).

This commit removes the script in debian/rules that detects the number of
cpus, requested by the environment variable DEBIAN_BUILD_NCPUS, which is not
official in debian.

Signed-off-by: Víctor Manuel Jáquez Leal <victorx.jaquez@intel.com>

https://bugzilla.gnome.org/show_bug.cgi?id=754087