Revert "dashdemux: Include the period start in the fragment timestamps in all cases"
This reverts commit e671ad25a989cb21c62c7a5867c2090890ce49ba.

The timestamps should restart at 0 again for each period, but we have to
adjust the segment to map those timestamps to the actual stream time and
running time of that period.

Otherwise we would have timestamps that conflict with the ones from the tfdt
inside the MP4 container, which are restarting at 0 for each period.

https://bugzilla.gnome.org/show_bug.cgi?id=752409