vaapisink: create display at open()
Instead of creating the VA display before setting the bus to the
element, it is created when the element is opened.

Basically, this commit is a revert of
5e5d62cac79754ba60057fc2516135aad8d7de35

That was done when the GStreamer's context sharing was not mature
enough as now. There is no reason to keep this hack.